feat: expand functionality of memory.Collection
Add constructor, using options pattern.
The options for memory.Collection have sensible defaults, so use the Functional Options pattern to set them.
There is a name collision, resulting in the awkward name `WithClockCollection`. This should be resolved in future.

Use clock.Clock as a source of timestamps.
The clock.Clock interface was moved into this repo. This is now the standard way to abstract a source of time, so use it instead of a function type.

Add accessor methods List, Get and AddFn.
These are required for changed to the electric trait memory implementation.