fix(wrap): always return an error from wrapped stream methods if the call failed
Methods Header, RecvMsg, SendMsg of both server and client streams should always return an error if the stream is closed. Previously if the stream was closed with a nil error, as it would if a server stream handler were to return nil, could cause the stream to get into a state where calling RecvMsg would fail but return a nil error, causing some applications (most) to go into an infinite loop!