resource: add WithMoreUpdateMask option for adding more paths
Before the process of adding to the list of update paths was a bit of a pain involving parsing the opts, checking the UpdateMask, adding to it, creating new opts to be parsed again. While people didn't need to do this frequently it is useful for things like presets.

Now you can just specify some new options to have extra update paths included.