wrap: fix deadlock when context is cancelled
If the context provided to a streaming call is cancelled, the sending and receiving sides raced on the closeErr.
The code falsely assumed that ClientServerStream.ctx ending implied that Close() had already been called,
but this might not be true if the parent context
was cancelled.
We must only read from closeErr (by calling closeErrLocked) if clientSend has been closed.