resource: fix collection change filter conversion
Add tests and fix the behaviour of the (*CollectionChange).include function. Before the change incoming adds of messages that shouldn't be included were returning ok=true indicating that they should be included in responses. This had the effect of Pull methods returning items that they shouldn't when those items were added after the pull started. We've noticed this in different apps but always put it down to incorrect app behaviour.

This change fixes these and adds 100% test coverage to the guilty function.